This pull request introduces changes to support floating-point values for the amount field in the distribution-related backend and frontend systems. The updates ensure consistency across the database schema, data models, and user interfaces, allowing more precise values for distribution amounts. Key changes include database schema migration, model updates, and frontend adjustments for better user experience and validation.

Backend Changes:
Database Schema Migration:

Altered the amount column in the distribution table from Integer to Float to support decimal values (04eb0b44ff7c).
Model Updates:

Updated the amount field in Distribution model (distribution_model.py) and create_distribution_dto to use Float instead of Integer. [1] [2]
Frontend Changes:
Form Enhancements:

Updated DistributionModal.vue to allow floating-point values for amount with a minimum of 0.1, a step of 0.25, and precision up to two decimal places. Added validation rules to enforce these constraints. [1] [2] [3]
UI Adjustments:

Modified the amount display in distributions.vue to show values with two decimal places.
Data Model Updates:

Adjusted the Distribution TypeScript interface to align with the backend changes, reflecting the new amount type and improving type definitions for related fields.

+++ Antonieta.Backend/models/distribution_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, ForeignKey, Integer
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, Float
 from sqlalchemy.orm import relationship
 from database import Base
 
@@ -8,7 +8,7 @@ class Distribution(Base):
     id = Column(Integer, primary_key=True, index=True)
     ration_id = Column(Integer, ForeignKey('ration_stock.id'), index=True)
     beneficiary_id = Column(Integer, ForeignKey('beneficiary.id'), index=True)
-    amount = Column(Integer, index=True)
+    amount = Column(Float, index=True)
     date = Column(DateTime, nullable=False)
     ration = relationship("RationStock", back_populates="distributions")
     beneficiary = relationship("Beneficiary", back_populates="distributions")
